Port over most of the tests from the original datalog implementation.
This adds some flexibility in terms of allowed characters, but does
not attempt to add the infix operators that the original allowed.

What's confusing is that the transitive closure test is failing "sometimes".
When failing, we recieve two duplicate answers q(a). The fact that it is
non-deterministic in this very single-threaded code makes me thing that
the fault lies in some map enumeration order.